feat(cmap): 添加基于代码点的字形ID获取方法并优化API
- 添加get_glyph_id_by_code方法，支持直接通过u32码点获取字形ID
- 重命名get_glyph_id方法注释，移除缓存优化相关描述
- 添加supported_chars方法，返回字体支持的字符列表
- 添加supported_chars_count方法，返回支持字符的数量
- 更新日志信息参数类型，使用code替代char_code

+++ crates/rfont/examples/font_info_demo.rs
@@ -41,7 +41,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     println!("━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━");
     let test_chars = vec!['阿', '里', '妈', '妈', 'A', 'B', '1', '2'];
     for ch in test_chars {
-        let supported = font.supports_character(ch as u32);
+        let supported = font.supports_character(ch);
         let status = if supported { "✅" } else { "❌" };
         println!("{} '{}' (U+{:04X})", status, ch, ch as u32);
     }
